fixup

+++ tests/openssh_fixture.c
@@ -413,7 +413,7 @@ static libssh2_socket_t open_socket_to_container(char *container_id)
     }
 
     sin.sin_family = AF_INET;
-    sin.sin_port = htons(atoi(port_string));
+    sin.sin_port = htons((unsigned short)atoi(port_string));
     sin.sin_addr.s_addr = hostaddr;
 
     for(counter = 0; counter < 3; ++counter) {
